Change handling of 5th week in recurrence modal
When editing a monthly recurrence, choosing the 5th instance of a
weekday in a month might result in an occurrence on the 4th instance of
that day, if in a month with only four such days. As a result, the
corresponding label on the week number select box has been changed from
'5th' to 'last'.